Divide summary component
