+++ src/components/Wallet/children/index.ts
@@ -1,4 +1,5 @@
 import { Accounts } from './Accounts/Accounts';
 import { NoExtension } from './NoExtension/NoExtension';
+import { Summary } from './Summary/Summary';
 
-export { Accounts, NoExtension };
+export { Accounts, NoExtension, Summary };
