+++ src/components/Wallet/hooks.ts
@@ -4,8 +4,6 @@ import {
   InjectedAccountWithMeta,
   InjectedExtension,
 } from '@polkadot/extension-inject/types';
-import { Balance } from '@polkadot/types/interfaces';
-import { useApi, useAccount } from 'hooks';
 
 export const useAccounts = () => {
   const [accounts, setAccounts] = useState<InjectedAccountWithMeta[]>();
@@ -20,21 +18,4 @@ export const useAccounts = () => {
   }, []);
 
   return accounts;
-};
-
-export const useBalance = () => {
-  const { api } = useApi();
-  const { account } = useAccount();
-  const [balance, setBalance] = useState<Balance>();
-
-  useEffect(() => {
-    if (account) {
-      const { address } = account;
-
-      api.balance.findOut(address).then(setBalance);
-      api.gearEvents.subscribeToBalanceChange(address, setBalance);
-    }
-  }, [api, account]);
-
-  return balance?.toHuman();
 };
